+++ __tests__/ord.e2e.test.js
@@ -370,4 +370,93 @@ describe("Tests for eventResource and apiResource", () => {
         expect(document.groups).toHaveLength(1);
         expect(document).toMatchSnapshot();
     });
+
+    it("should block OpenResourceDiscoveryService", async () => {
+        const linkedModel = cds.linked(`
+                service OpenResourceDiscoveryService {
+                    event ServiceEvent : {
+                        ID    : Integer;
+                    }
+                };
+                service MyService {
+                    event ServiceEvent : {
+                        ID    : Integer;
+                    }
+                    function getRatings() returns Integer;
+                };
+                annotate MyService with @ORD.Extensions : {
+                    visibility : 'internal'
+                };
+            `);
+
+        const document = ord(linkedModel);
+        expect(document.apiResources).toHaveLength(1);
+        expect(document.eventResources).toHaveLength(1);
+        expect(document.groups).toHaveLength(1);
+        expect(document).toMatchSnapshot();
+    });
+
+    it("should exclude service when it has procotol none", async () => {
+        const linkedModel = cds.linked(`
+                @protocol: 'none'
+                service InternalService {
+                    event InternalEvent: {
+                        ID: Integer;
+                    }
+
+                    entity InternalEntity {
+                        ID: Integer;
+                    }
+                }
+
+                service MyService {
+                    event ServiceEvent : {
+                        ID    : Integer;
+                    }
+                    function getRatings() returns Integer;
+                };
+                annotate MyService with @ORD.Extensions : {
+                    visibility : 'internal'
+                };
+            `);
+
+        const document = ord(linkedModel);
+        expect(document.apiResources).toHaveLength(1);
+        expect(document.eventResources).toHaveLength(1);
+        expect(document.groups).toHaveLength(1);
+        expect(document).toMatchSnapshot();
+    });
+
+    it("should exclude external service", async () => {
+        const linkedModel = cds.linked(`
+                service ExternalService {
+                    event InternalEvent: {
+                        ID: Integer;
+                    }
+
+                    entity InternalEntity {
+                        ID: Integer;
+                    }
+                }
+
+                service MyService {
+                    event ServiceEvent : {
+                        ID    : Integer;
+                    }
+                    function getRatings() returns Integer;
+                };
+                annotate MyService with @ORD.Extensions : {
+                    visibility : 'internal'
+                };
+            `);
+        cds.requires.ExternalService = {
+            kind: "external",
+            model: "external-model",
+        };
+        const document = ord(linkedModel);
+        expect(document.apiResources).toHaveLength(1);
+        expect(document.eventResources).toHaveLength(1);
+        expect(document.groups).toHaveLength(1);
+        expect(document).toMatchSnapshot();
+    });
 });