+++ CHANGELOG.md
@@ -4,7 +4,72 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
-## 1.3.3 (2024-06-25)
+## 1.3.9 (2025-09-02)
+
+## What's Changed
+
+- fix: If the visibility is private, do not create a group by @KegCoder
+- feat: allow loading of additional attributes for packages
+
+## New Contributors
+
+- @KegCoder made their first contribution in https://github.com/cap-js/ord/pull/215
+
+**Full Changelog**: https://github.com/cap-js/ord/compare/v1.3.8...v1.3.9
+
+## 1.3.8 (2025-09-01)
+
+## What's Changed
+
+- Feat: Setup Java sample
+- Feat: setup java pipeline
+- Fix: Support CommonJs
+
+**Full Changelog**: https://github.com/cap-js/ord/compare/v1.3.7...v1.3.8
+
+## 1.3.7 (2025-08-19)
+
+## What's Changed
+
+- Feat: use bcryptjs
+
+## 1.3.6 (2025-08-07)
+
+## What's Changed
+
+- Feat: add progressBar
+- Feat: support exposedEntityTypes
+- Feat: introducing defaultVisibility parameter in config
+- Feat: skip ord service for java project
+
+## New Contributors
+
+- @SAPzer0 made their first contribution in https://github.com/cap-js/ord/pull/206
+
+**Full Changelog**: https://github.com/cap-js/ord/compare/v1.3.5...v1.3.6
+
+## 1.3.5 (2025-07-17)
+
+## What's Changed
+
+- Block OpenResourceDiscoveryService in ord document
+
+**Full Changelog**: https://github.com/cap-js/ord/compare/v1.3.4...v1.3.5
+
+## 1.3.4 (2025-07-09)
+
+## What's Changed
+
+- Adjust default values for vendor to "customer:vendor:Customer"
+- Improved runtime of cmd `cds build --for ord`
+- Alignment of publishing with SAP Open Resource Discovery (ORD) global repository strategy
+- Fix: DP package assignment
+- Fix: error of ORD-IDs with duplicate namespace components
+- Fix: error of cmd `cds build --for ord`
+
+**Full Changelog**: https://github.com/cap-js/ord/compare/v1.3.4...v1.3.3
+
+## 1.3.3 (2025-06-25)
 
 ## What's Changed
 
@@ -13,7 +78,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 **Full Changelog**: https://github.com/cap-js/ord/compare/v1.3.2...v1.3.3
 
-## 1.3.2 (2024-06-18)
+## 1.3.2 (2025-06-18)
 
 ## What's Changed
 
@@ -21,7 +86,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 **Full Changelog**: https://github.com/cap-js/ord/compare/v1.3.1...v1.3.2
 
-## 1.3.1 (2024-06-13)
+## 1.3.1 (2025-06-13)
 
 ### What's Changed
 
@@ -31,7 +96,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 **Full Changelog**: https://github.com/cap-js/ord/compare/v1.3.0...v1.3.1
 
-## 1.3.0 (2024-05-20)
+## 1.3.0 (2025-05-20)
 
 ### What's Changed